+++ CHANGELOG.rst
@@ -7,8 +7,12 @@ TBR
 
 * added a ``PageObjectRegistry`` class which has the  ``handle_urls`` decorator
   to write override rules.
-* removed support for Python 3.6
-* added support for Python 3.10
+* Removed support for Python 3.6
+* Added support for Python 3.10
+* Added support for performing additional requests using
+  ``web_poet.HttpClient``.
+* Introduced ``web_poet.Meta`` to pass arbitrary information
+  inside a Page Object.
 
 0.1.1 (2021-06-02)
 ------------------